separate schema objects into files
